Flatten mysql until necessary
